Add files via upload
The purpose of this project is to modify an already existing network architecture and change its configuration to fit a new set of specifications. This is achieved by following a step-by-step tutorial presented in the class’s content, which also includes the already existing topology that will be used to meet the new set of requirements. An already existing packet tracer file will serve as the starting point for the tutorial, where we will be modifying the IP addresses from a class C to a class A in order to create increased IP space. 
First, we start off with the configuration of the interfaces, this is done based off the IP table presented in the tutorial, in this section we will be skipping the f0/0.10-40 set up on R1. We will be setting up Outer, R1 (with the exception of the specified interfaces), S1, and PC6 which will be the management device. We will be following a set of steps that can be observed below.
1.	In the Cli of each device insert the following commands: conf t > int [interface indicated in the given table for the specific device] > ip add [IP address and mask] > no shutdown > end. 
2.	Then, on S1 we will configure the Vlans as directed conf t > vlan [vlan number] > name [indicated name]. Vlan 10 is Office, Vlan 20 is Public, Vlan 30 is Corporate, and Vlan 40 is Management. Then command “exit” to finish. Following the same steps create a Vlan 999 with name Black_Hole, then command “end”. This is assigned to ports that are not currently in use, it’s done for best practice.
3.	Still on S1, follow the next commands: conf t > int ran f0/8-23,g0/1-2 > shutdown > switchport mode access > switchport access vlan 999. These steps are to disable unused interfaces and assign them to the Black_Hole Vlan. 
4.	Hosts 1-7 (still on S1) are configured as follows: int ran f0/1-7 > switchport mode access > switchport port-security > switchport port-security mac-address sticky. Then assign the corresponding Vlans to their corresponding ports int ran [select ports in ranges of 2, e.g. “f0/1-2”] > switchport access vlan [insert the vlan number to be assigned]. In this case the corresponding ports to Vlans go as specified: Ports 1-2 correspond to Vlan 10, and so on until reaching ports 5-6 (Vlan 30), port 7 is assigned to Vlan 40 alone. 
5.	For interface f0/24 command switchport mode trunk > switchport nonegotiate > switchport trunk allowed vlan 10,20,30,40 > switchport trunk native vlan 40. This is done to configure a trunk passing through the router, as well as native vlan 40. Exit S1.
6.	Now, to configure R1 we will create 4 subinterfaces by following the commands: enable > conf t > int g0/0.10 > encapsulation dot1q 10 > ip add 10.10.0.1 255.255.255.0. Repeat for subinterfaces g0/0.20-30, replacing the underlined number on the IP address that corresponds to said subinterface. For g0/0.40 follow the same steps, but in encapsulation add “native” after entering the associated number. Then command “exit” > int g0/0 > no shutdown > end. R1 should look as showed in the screenshot below:
 
7.	Now, we will configure the router as a dhcp server. Conf t > ip dhcp excluded address 10.10.0.1. Then follow the same steps, but replace the underlined number with 20-30. The following steps are to create the dhcp pools: ip dhcp pool VLAN_10 > default-router 10.10.0.1 > network 10.10.0.0 255.255.255.0 > exit. Follow the same steps for vlans 20-30, replacing the underlined numbers for the corresponding ones. 
8.	Now we will create an access list to prevent vlans 10-30 from accessing the management pc, without taking access away from management to vlans 10-30. Conf t > ip access-list extended MANAGEMENT_FILTER > permit ip 192.168.40.0 0.0.0.7 any > permit icmp 192.168.40.0 0.0.0.7 any > deny ip any any > deny icmp any any > exit > int g0/0.40 > ip access-group MANAGEMENT_FILTER out > end. 
9.	To block other vlans from accessing R1/S1 follow the steps: conf t > ip access-list extended KEEP_OUT > permit tcp host 192.168.40.3 any eq 22 > exit > line vty 0 4 > access-class KEEP_OUT in > end.
10.	Next, we will do routing on R1 and Outer devices. On R1 conf t > router eigrp > eigrp router-id 1.1.1.1 > no auto > passive-interface g0/0 [then insert passive-interface for g0/0.10-40] > net 10.10.0.0 0.0.0.255 [then assign networks 20-40 by replacing the underlined number] > end. Follow the same steps for Outer, the router-id is 2.2.2.2, the only passive-interface is lo0, and the networks to be added are 172.16.0.0 0.0.0.3 and 209.165.200.1 0.0.0.0. 
11.	For static routing to Outer device from R1 conf t > ip route 0.0.0.0 0.0.0.0 172.16.0.1. On Outer conf t > ip route 10.10.0.0 255.255.255.0 172.16.0.2 [replace underlined number 20-30]. For lo0 as follows: ip route 0.0.0.0 0.0.0.0 lo0 4 > end.
We have reached the end of the tutorial, now we can run some pings from the end devices. In this case, vlans 10-30 should be able to ping each other, but not the management device (vlan 40). However, the management device should be able to ping all vlans. After setting up each end device to a dhcp gateway the default gateway and IP address should change to one belonging to an IP class A address within the vlan. For example, PC0 that belongs to Vlan 10 has an IP address of 10.10.0.3, and so on for each device within each Vlan. It is important to point out that we should give the packet tracer (finished) file some time to load in order to see the right IP addresses pop up.
In conclusion, we have been able to utilize the preset packet tracer file and follow each step in the tutorial provided to set up our small network. The details that were changed in comparison to the original tutorial include the IP address space change, which we changed from class C IP addresses to class A ones. With this change we have created space for over 64,000 possible IP addresses per network. The only exception to this modification is the Management PC (Vlan 40) which stayed the same as the example provided (as desired).