Add files via upload
This lab includes the configuration of an existing topology. IP's were changed from class C to class A to create more space, VLANs 10-30 are able to communicate with each other while VLAN 40 (management) is blocked off. VLAN 40 on the other hand is able to ping all VLANs in the set up.